Bundle includes compiled SASS
* Install node-sass and sass-loader, css-loader, style-loader.
* Add loaders to webpack config
* Include main.sass in entry point main.js
* In dev environment, node modules may be installed from a different
architecture than the Docker image. Need to install linux node-sass
binary after npm install to ensure that docker-compose will have the
correct binary. This isn't a problem for the production docker
container because the npm install will happen in the linux
environment and so it will already have the correct image.