Automatically save message when changing to new selected message
This causes a problem because it changes the message name but does
not update the name in the list. Should we allow message name to
change? Should we be smart and update any exits? Right now changing
the name creates a new entry in the db.

Also does not provide any user feedback. Errors will be incorrectly
reported. Rather than storing the name of the message being saved
and the error message, maybe we should store a list of saving
messages and a list of errors, indexed by message name. That would
make it easier to provide feedback in the message list and if you
select a particular message.

* Use snapshot to save contents of selected message before update
* If selected message has changed and the previous message changed
the data from what it got passed in as props, save the previous
message.