GlobalBlot can be inserted using the button in the toolbar
Add GlobalBlot inline embed type that can drag/drop in the editor
and be inserted through clicking the button in the toolbar.

GlobalEmbed is the generic type, but things right now say FirstName
because that's the first thing I want to implement when parsing
blots and it's not handling modifying the global yet.

* Toolbar loaded from component
* Replace old toolbar with one that has a button to insert the
blot
* Move argie-mod to .tsx because it uses jsx for rendering global
embed