Clicking GetGlobalBlot allows editing the globalName
* React component controls globalName and editable data in node and in
React props/state so move initialization into React component. Component
will create node and set all relevant properties on the node and then
modify the properties as the data changes. The Component essentially
controls the blot's data. The GetGlobalBlot class handles creating the
node, rendering the component and passing it the initial data and node,
and then the required and necessary static methods and properties as
well as the static methods for going from delta to text and vice versa.

* When an editable GetGlobalBlot is clicked, editing state is true
and text is replaced by input. When editing is finished, editing
state is false and the node's dataset is updated with the new
globalName. GetGlobalBlot class uses the dataset to determine the
blot's value and thus correctly converts to text with the updated
globalName value.

One problem is that editable is hardcoded to globalName, so setting
a custom GetGlobalBlot to lastName/firstName will make the blot
uneditable after translating to text and back. A possible solution
is to include a second parameter to avatarWrapper.getGlobal() that
is ignored but stores editable.